apps: rename ok test to internaltest
