+++ hack/import-restrictions.json
@@ -471,7 +471,7 @@
       "github.com/openshift/origin/pkg/api/meta",
       "github.com/openshift/origin/pkg/apps/apis/apps",
       "github.com/openshift/origin/pkg/apps/apis/apps/install",
-      "github.com/openshift/origin/pkg/apps/apis/apps/test",
+      "github.com/openshift/origin/pkg/apps/apis/apps/internaltest",
       "github.com/openshift/origin/pkg/apps/client/internalversion",
       "github.com/openshift/origin/pkg/apps/client/v1",
       "github.com/openshift/origin/pkg/apps/controller/util",
