+++ pkg/apps/controller/util/util_test.go
@@ -9,7 +9,7 @@ import (
 	kapi "k8s.io/kubernetes/pkg/apis/core"
 
 	appsapi "github.com/openshift/origin/pkg/apps/apis/apps"
-	appstest "github.com/openshift/origin/pkg/apps/apis/apps/test"
+	appstest "github.com/openshift/origin/pkg/apps/apis/apps/internaltest"
 
 	_ "github.com/openshift/origin/pkg/api/install"
 )
